+++ CMakeLists.txt
@@ -30,6 +30,8 @@ include(CTest)
 
 add_executable(trie_test ${CMAKE_CURRENT_SOURCE_DIR}/tests/trie_test.c)
 add_executable(ip_test ${CMAKE_CURRENT_SOURCE_DIR}/tests/ip_test.c)
+add_executable(tool_test ${CMAKE_CURRENT_SOURCE_DIR}/tests/tool_test.c)
+add_executable(ip_test_not_init ${CMAKE_CURRENT_SOURCE_DIR}/tests/ip_test_not_initialized.c)
 
 
 target_link_libraries(trie_test PUBLIC
@@ -42,8 +44,20 @@ target_link_libraries(ip_test PUBLIC
         unity
 )
 
+target_link_libraries(tool_test PUBLIC
+        ip_trie_lib
+        unity
+)
+
+target_link_libraries(ip_test_not_init PUBLIC
+        ip_trie_lib
+        unity
+)
+
 add_test(NAME trie_test COMMAND trie_test)
 add_test(NAME ip_test COMMAND ip_test)
+add_test(NAME tool_test COMMAND tool_test)
+add_test(NAME ip_test_not_init COMMAND ip_test_not_init)
 
 add_custom_target(coverage
         COMMAND lcov --capture --directory . --output-file coverage.info
@@ -55,10 +69,6 @@ add_custom_target(cccc
         COMMAND cccc src/*.c include/*.h
         WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
 
-add_custom_target(pdm
-        COMMAND pmd-bin-6.34.0/bin/run.sh cpd --files ${CMAKE_CURRENT_SOURCE_DIR} --format xml --language cpp --minimum-tokens 20
-        WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
-
 add_custom_target(cppcheck
         COMMAND cppcheck --enable=all --inconclusive --xml --xml-version=2 --force --language=c --std=c23 --platform=unix64
         --output-file=${CMAKE_CURRENT_BINARY_DIR}/cppcheck.xml ${CMAKE_CURRENT_SOURCE_DIR}/src ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_SOURCE_DIR}/tests
@@ -66,6 +76,14 @@ add_custom_target(cppcheck
         WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
 
 add_custom_target(valgrind
-        COMMAND valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes ${CMAKE_CURRENT_BINARY_DIR}/trie_test  > /dev/null
-        COMMAND valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes ${CMAKE_CURRENT_BINARY_DIR}/ip_test  > /dev/null
+        COMMAND valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes ${CMAKE_CURRENT_BINARY_DIR}/trie_test > /dev/null
+        COMMAND valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes ${CMAKE_CURRENT_BINARY_DIR}/ip_test > /dev/null
         WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
+
+add_custom_target(check
+        COMMAND ${CMAKE_CTEST_COMMAND} --output-on-failure
+        DEPENDS trie_test
+        ip_test
+        tool_test
+        ip_test_not_init
+)
